remind_me_later/requirements.txt
Django==4.2.6
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv 
drf-yasg
djangorestframework-simplejwt==4.7.2
drf-yasg==1.21.5
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
#asgiref==3.7.2
django-cors-headers==4.2.0
bcrypt==4.0.1
setuptools
coreapi